Move packaging code into Editor folder
