--- Assets/Digital Painting/Scripts/DevTest/PackageBuilder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 933818cbc772cfe46bd90f667f927c00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
