+++ Assets/Digital Painting/Scripts/Editor/PackageBuilder.cs
@@ -12,67 +12,39 @@ using UnityEngine;
 /// Run from the command line with:
 ///   `"C:\Program Files\Unity\Editor\Unity.exe" -executeMethod PackageBuilder.Build`
 /// </summary>
-public class PackageBuilder
-{
+public class PackageBuilder {
 
     [MenuItem("Digital Painting/Build/Build Core Package")]
     public static void Build()
     {
-        string rootDir = "Assets\\Digital Painting";
+        string rootDir = "Assets/Digital Painting";
         string excludeSubDir = "Plugins";
         string packageName = "DigitalPainting.unitypackage";
 
         // Delete everything in plugins directory except *.unitypackage and *.md (and matching .meta)
-        MoveExcludedFiles(rootDir + "\\" + excludeSubDir);
+        RemoveDirs(rootDir + "/" + excludeSubDir);
         AssetDatabase.Refresh();
 
         AssetDatabase.ExportPackage(rootDir, packageName, ExportPackageOptions.Interactive | ExportPackageOptions.Recurse);
         Debug.Log("Exported " + packageName);
-
-        RecoverExcludedFiles(rootDir + "\\" + excludeSubDir);
-        AssetDatabase.Refresh();
     }
 
-    protected static void MoveExcludedFiles(string dir)
+    private static void RemoveDirs(string dir)
     {
         string[] subdirectoryEntries = Directory.GetDirectories(dir);
         foreach (string subdirectory in subdirectoryEntries)
         {
             if (Path.GetFileName(subdirectory) != "Scenes")
             {
-                Debug.Log("Moving to safety: " + subdirectory);
-                string copyPath = "Temp" + Path.DirectorySeparatorChar + subdirectory;
-                Directory.CreateDirectory(Path.GetDirectoryName(copyPath));
-                Directory.Move(subdirectory, copyPath);
-                File.Move(subdirectory + ".meta", copyPath + ".meta");
+                Debug.Log("Deleting " + subdirectory + " and associated `.meta` file.");
+                Directory.Delete(subdirectory, true);
+                File.Delete(subdirectory + ".meta");
             }
             else
             {
-                MoveExcludedFiles(subdirectory);
+                RemoveDirs(subdirectory);
             }
         }
-    }
-
-    protected static void RecoverExcludedFiles(string dir)
-    {
-        string copyPath = "Temp" + Path.DirectorySeparatorChar + dir;
-        string[] subdirectoryEntries = Directory.GetDirectories(copyPath);
-        foreach (string subdirectory in subdirectoryEntries)
-        {
-            if (Path.GetFileName(subdirectory) != "Scenes")
-            {
-                string targetPath = subdirectory.Substring(subdirectory.IndexOf(Path.DirectorySeparatorChar, 1) + 1);
-                Debug.Log("Moving back to project from: " + subdirectory + " to " + targetPath);
-                Directory.CreateDirectory(Path.GetDirectoryName(targetPath));
-                Directory.Move(subdirectory, targetPath);
-                File.Move(subdirectory + ".meta", targetPath + ".meta");
-            }
-            else
-            {
-                MoveExcludedFiles(subdirectory);
-            }
-        }
-        Directory.Delete(copyPath, true);
     }
 
 }
